+++ modules/apps/headless/headless-builder/headless-builder-test/src/testIntegration/java/com/liferay/headless/builder/model/listener/test/APISchemaObjectEntryModelListenerTest.java
@@ -12,10 +12,9 @@
  * details.
  */
 
-package com.liferay.headless.builder.test.validation.test;
+package com.liferay.headless.builder.model.listener.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
@@ -36,7 +35,7 @@ import org.junit.runner.RunWith;
  */
 @Ignore
 @RunWith(Arquillian.class)
-public class APIApplicationObjectEntryModelListenerTest {
+public class APISchemaObjectEntryModelListenerTest {
 
 	@ClassRule
 	@Rule
@@ -44,27 +43,26 @@ public class APIApplicationObjectEntryModelListenerTest {
 		new LiferayIntegrationTestRule();
 
 	@Test
-	public void testInvalidBaseURLPathAPIApplication() throws Exception {
+	public void testPostAPISchemaNotRelatedWithAPIApplication()
+		throws Exception {
+
 		JSONObject jsonObject = HTTPTestUtil.invoke(
 			JSONUtil.put(
-				"applicationStatus", "published"
+				"mainObjectDefinitionERC", RandomTestUtil.randomString()
 			).put(
-				"baseURL",
-				RandomTestUtil.randomString() + StringPool.FORWARD_SLASH
-			).put(
-				"title", RandomTestUtil.randomString()
+				"name", RandomTestUtil.randomString()
 			).toString(),
-			"headless-builder/applications", Http.Method.POST);
+			"headless-builder/schemas", Http.Method.POST);
 
 		Assert.assertEquals("BAD_REQUEST", jsonObject.get("status"));
 		Assert.assertEquals(
-			"Base URL can have a maximum of 255 alphanumeric characters",
+			"An schema must be related to an application",
 			jsonObject.get("title"));
 	}
 
 	@Test
-	public void testValidBaseURLPathAPIApplication() throws Exception {
-		JSONObject jsonObject = HTTPTestUtil.invoke(
+	public void testPostAPISchemaRelatedWithAPIApplication() throws Exception {
+		JSONObject apiApplicationJSONObject = HTTPTestUtil.invoke(
 			JSONUtil.put(
 				"applicationStatus", "published"
 			).put(
@@ -73,6 +71,17 @@ public class APIApplicationObjectEntryModelListenerTest {
 				"title", RandomTestUtil.randomString()
 			).toString(),
 			"headless-builder/applications", Http.Method.POST);
+
+		JSONObject jsonObject = HTTPTestUtil.invoke(
+			JSONUtil.put(
+				"mainObjectDefinitionERC", RandomTestUtil.randomString()
+			).put(
+				"name", RandomTestUtil.randomString()
+			).put(
+				"r_apiApplicationToAPISchemas_c_apiApplicationId",
+				apiApplicationJSONObject.getLong("id")
+			).toString(),
+			"headless-builder/schemas", Http.Method.POST);
 
 		Assert.assertEquals(
 			0,
